Refactor more existing code to use `AppendCommaSeparated`
