+++ tests/Moq.Tests/ExtensionsFixture.cs
@@ -103,6 +103,37 @@ namespace Moq.Tests
 
 			Assert.False(thisMethod.IsExtensionMethod());
 		}
+
+		[Theory]
+		[InlineData(nameof(IMethods.Empty), "")]
+		[InlineData(nameof(IMethods.Int), "int")]
+		[InlineData(nameof(IMethods.IntAndString), "int, string")]
+		[InlineData(nameof(IMethods.InInt), "in int")]
+		[InlineData(nameof(IMethods.RefInt), "ref int")]
+		[InlineData(nameof(IMethods.OutInt), "out int")]
+		[InlineData(nameof(IMethods.BoolAndParamsString), "bool, params string[]")]
+		public void GetParameterTypeList_formats_parameter_lists_correctly(string methodName, string expected)
+		{
+			var actual = GetParameterTypeList(methodName);
+			Assert.Equal(expected, actual);
+		}
+
+		private string GetParameterTypeList(string methodName)
+		{
+			var method = typeof(IMethods).GetMethod(methodName);
+			return method.GetParameterTypeList();
+		}
+
+		public interface IMethods
+		{
+			void Empty();
+			void Int(int arg1);
+			void IntAndString(int arg1, string arg2);
+			void InInt(in int arg1);
+			void RefInt(ref int arg1);
+			void OutInt(out int arg1);
+			void BoolAndParamsString(bool arg1, params string[] arg2);
+		}
 	}
 
 	public interface IFooReset
